pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "elelem"
dynamic = ["version"]
description = "Unified API wrapper for OpenAI, GROQ, and DeepInfra with cost tracking and retry logic"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Fable Team", email = "dev@fable.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.99.2",
    "groq>=0.4.0",
    "aiofiles>=23.0.0",
    "jsonschema>=4.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.1",
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.33.0",
    "pydantic>=2.10.6",
    "pandas>=2.0.3",
    "requests>=2.32.4",
    "pytest>=8.3.5",
    "flask>=3.0.3",
    "pytest-asyncio>=0.24.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourorg/elelem"
Repository = "https://github.com/yourorg/elelem"
Documentation = "https://github.com/yourorg/elelem/blob/main/README.md"
Issues = "https://github.com/yourorg/elelem/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
elelem = ["config.json", "models.yaml"]

[tool.setuptools_scm]
write_to = "src/elelem/_version.py"
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
fallback_version = "0.1.0.dev0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
